+++ activerecord/lib/active_record/associations/builder/association.rb
@@ -51,5 +51,19 @@ module ActiveRecord::Associations::Builder
           association(name).writer(value)
         end
       end
-  end
+
+      def dependent_restrict_raises?
+        ActiveRecord::Base.dependent_restrict_raises == true
+      end
+
+      def dependent_restrict_deprecation_warning
+        if dependent_restrict_raises?
+          msg = "In the next release, `:dependent => :restrict` will not raise a `DeleteRestrictionError`."\
+                "Instead, it will add an error on the model. To fix this warning, make sure your code" \
+                "isn't relying on a `DeleteRestrictionError` and then add" \
+                "`config.active_record.dependent_restrict_raises = false` to your application config."
+          ActiveSupport::Deprecation.warn msg
+        end
+      end
+   end
 end